refactor: split common utility modules
Extract logging, dataframe, subprocess, HTTP, and error helpers
from common.utils into focused common modules. Keep compatibility
re-exports and thin wrappers in common.utils so existing call sites
and tests continue to work during the transition.

Update representative consumers to import the new modules directly,
reducing the blast radius of future changes in vulnxscan and the
remaining common.utils call sites.
