+++ src/common/log.py
@@ -1,0 +1,56 @@
+# SPDX-FileCopyrightText: 2026 Technology Innovation Institute (TII)
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""Shared logging configuration and logger access."""
+
+import logging
+import os
+
+from colorlog import ColoredFormatter, default_log_colors
+
+LOG_SPAM = logging.DEBUG - 1
+LOG = logging.getLogger(os.path.abspath(__file__))
+
+
+def set_log_verbosity(verbosity=1):
+    """Set logging verbosity."""
+    log_levels = [logging.NOTSET, logging.INFO, logging.DEBUG, LOG_SPAM]
+    verbosity = min(len(log_levels) - 1, max(verbosity, 0))
+    _init_logging(verbosity)
+
+
+def _init_logging(verbosity=1):
+    """Initialize logging."""
+    if verbosity == 0:
+        level = logging.NOTSET
+    elif verbosity == 1:
+        level = logging.INFO
+    elif verbosity == 2:
+        level = logging.DEBUG
+    else:
+        level = LOG_SPAM
+    if level <= logging.DEBUG:
+        logformat = (
+            "%(log_color)s%(levelname)-8s%(reset)s "
+            "%(filename)s:%(funcName)s():%(lineno)d "
+            "%(message)s"
+        )
+    else:
+        logformat = "%(log_color)s%(levelname)-8s%(reset)s %(message)s"
+    logging.addLevelName(LOG_SPAM, "SPAM")
+    default_log_colors["INFO"] = "fg_bold_white"
+    default_log_colors["DEBUG"] = "fg_bold_white"
+    default_log_colors["SPAM"] = "fg_bold_white"
+    formatter = ColoredFormatter(logformat, log_colors=default_log_colors)
+    if LOG.hasHandlers() and len(LOG.handlers) > 0:
+        stream = LOG.handlers[0]
+    else:
+        stream = logging.StreamHandler()
+    stream.setFormatter(formatter)
+    if not LOG.hasHandlers():
+        LOG.addHandler(stream)
+    LOG.setLevel(level)
+
+
+set_log_verbosity(1)
