+++ src/repology/repology_cli.py
@@ -16,7 +16,8 @@ import pandas as pd
 from tabulate import tabulate
 
 import repology.exceptions
-from common.utils import LOG, df_to_csv_file, set_log_verbosity
+from common.df import df_to_csv_file
+from common.log import LOG, set_log_verbosity
 from repology.adapter import RepologyAdapter, RepologyQuery
 
 ###############################################################################
