+++ src/repology/repology_cve.py
@@ -19,7 +19,9 @@ from bs4 import BeautifulSoup
 from tabulate import tabulate
 
 import repology.exceptions
-from common.utils import LOG, LOG_SPAM, df_to_csv_file, parse_version, set_log_verbosity
+from common.df import df_to_csv_file
+from common.log import LOG, LOG_SPAM, set_log_verbosity
+from common.utils import parse_version
 from repology.session import DEFAULT_REPOLOGY_SESSION, REPOLOGY_REQUEST_TIMEOUT
 
 ###############################################################################
