+++ src/sbomnix/derivation.py
@@ -14,8 +14,9 @@ import json
 
 from packageurl import PackageURL
 
+from common.log import LOG, LOG_SPAM
 from common.nix_utils import parse_nix_derivation_show
-from common.utils import LOG, LOG_SPAM, exec_cmd, nix_cmd
+from common.proc import exec_cmd, nix_cmd
 
 ###############################################################################
